--- app/src/main/java/uk/ac/tees/donut/squad/HomeFragment.java
@@ -1,19 +1,0 @@
-package uk.ac.tees.donut.squad;
-
-/**
- * Created by jlc-1 on 31/03/2017.
- */
-
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-public class HomeFragment extends Fragment
-{
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.home_fragment, container, false);
-    }
-}
