+++ app/src/main/java/uk/ac/tees/donut/squad/activities/MenuActivity.java
@@ -1,0 +1,57 @@
+package uk.ac.tees.donut.squad.activities;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+
+import uk.ac.tees.donut.squad.R;
+import uk.ac.tees.donut.squad.location.LocationActivity;
+import uk.ac.tees.donut.squad.location.MapActivity;
+
+public class MenuActivity extends AppCompatActivity {
+    boolean firstStart = true;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+
+    //Click functionality
+    public void openProfile(View view) {
+        Intent intent = new Intent(this, ProfileActivity.class);
+        startActivity(intent);
+    }
+
+    public void openSquads(View view) {
+        Intent intent = new Intent(this, SquadsActivity.class);
+        startActivity(intent);
+    }
+
+    public void openEvents(View view) {
+        Intent intent = new Intent(this, ViewMeetups.class);
+        startActivity(intent);
+    }
+
+    public void openHost(View view) {
+        Intent intent = new Intent(this, NewMeetup.class);
+        startActivity(intent);
+    }
+
+    public void openSettings(View view) {
+        Intent intent = new Intent(this, SettingsActivity.class);
+        startActivity(intent);
+    }
+
+    public void openLocation(View view) {
+    Intent intent = new Intent (this, LocationActivity.class);
+        startActivity(intent);
+
+    }
+
+    public void openMap(View view){
+        Intent intent = new Intent(this, MapActivity.class);
+        startActivity(intent);
+    }
+}
