+++ app/src/main/java/uk/ac/tees/donut/squad/activities/NewMeetup.java
@@ -1,4 +1,4 @@
-package uk.ac.tees.donut.squad;
+package uk.ac.tees.donut.squad.activities;
 
 import android.content.DialogInterface;
 import android.support.annotation.NonNull;
@@ -25,6 +25,7 @@ import com.google.firebase.database.ValueEventListener;
 import java.util.ArrayList;
 import java.util.List;
 
+import uk.ac.tees.donut.squad.R;
 import uk.ac.tees.donut.squad.posts.Meetup;
 
 public class NewMeetup extends AppCompatActivity
