+++ app/src/main/java/uk/ac/tees/donut/squad/activities/SplashScreen.java
@@ -1,15 +1,11 @@
-package uk.ac.tees.donut.squad;
+package uk.ac.tees.donut.squad.activities;
 
 import android.os.Bundle;
 import android.content.Intent;
-import android.app.Activity;
-import android.os.Handler;
 import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
-import android.widget.Button;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import com.google.android.gms.auth.api.Auth;
@@ -25,6 +21,8 @@ import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.auth.GoogleAuthProvider;
+
+import uk.ac.tees.donut.squad.R;
 
 /**
  * Created by jlc-1 on 21/03/2017.
