+++ app/src/main/java/uk/ac/tees/donut/squad/activities/SquadsActivity.java
@@ -1,7 +1,9 @@
-package uk.ac.tees.donut.squad;
+package uk.ac.tees.donut.squad.activities;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import uk.ac.tees.donut.squad.R;
 
 public class SquadsActivity extends AppCompatActivity {
