+++ app/src/main/java/uk/ac/tees/donut/squad/activities/fragments/MeetupsFragment.java
@@ -1,17 +1,20 @@
-package uk.ac.tees.donut.squad;
+package uk.ac.tees.donut.squad.activities.fragments;
 
 /**
  * Created by jlc-1 on 31/03/2017.
  */
+
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class ProfileFragment extends Fragment {
+import uk.ac.tees.donut.squad.R;
+
+public class MeetupsFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.profile_fragment, container, false);
+        return inflater.inflate(R.layout.meetups_fragment, container, false);
     }
 }
