+++ app/src/main/java/uk/ac/tees/donut/squad/activities/fragments/SquadsFragment.java
@@ -1,4 +1,4 @@
-package uk.ac.tees.donut.squad;
+package uk.ac.tees.donut.squad.activities.fragments;
 
 /**
  * Created by jlc-1 on 31/03/2017.
@@ -10,9 +10,12 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class MeetupsFragment extends Fragment {
+import uk.ac.tees.donut.squad.R;
+
+public class SquadsFragment extends Fragment {
+
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.meetups_fragment, container, false);
+        return inflater.inflate(R.layout.squads_fragment, container, false);
     }
 }
