+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/activity_main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="uk.ac.tees.donut.squad.MainActivity"
+    tools:context="uk.ac.tees.donut.squad.activities.MainActivity"
     android:background="@color/cardview_dark_background"
     >
