+++ app/src/main/res/layout/activity_meetup_detail.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="uk.ac.tees.donut.squad.MeetupDetail"
+    tools:context="uk.ac.tees.donut.squad.activities.MeetupDetail"
     android:background="@color/cardview_dark_background">
 
     <RelativeLayout
