+++ app/src/main/res/layout/activity_new_meetup.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="uk.ac.tees.donut.squad.NewMeetup"
+    tools:context="uk.ac.tees.donut.squad.activities.NewMeetup"
     android:background="@color/cardview_dark_background">
 
     <RelativeLayout
