+++ app/src/main/res/layout/activity_settings.xml
@@ -8,6 +8,6 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="uk.ac.tees.donut.squad.SettingsActivity">
+    tools:context="uk.ac.tees.donut.squad.activities.SettingsActivity">
 
 </RelativeLayout>
