+++ app/src/main/res/layout/activity_squads.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="uk.ac.tees.donut.squad.SquadsActivity">
+    tools:context="uk.ac.tees.donut.squad.activities.SquadsActivity">
 
     <TextView
         android:text="Squads:"
