+++ app/src/main/res/layout/activity_view_meetups.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="uk.ac.tees.donut.squad.ViewMeetups">
+    tools:context="uk.ac.tees.donut.squad.activities.ViewMeetups">
 
     <android.support.v7.widget.RecyclerView
         android:layout_width="match_parent"
