+++ app/src/main/res/layout/profile_fragment.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="uk.ac.tees.donut.squad.ProfileActivity"
+    tools:context="uk.ac.tees.donut.squad.activities.ProfileActivity"
     android:background="@color/cardview_dark_background">
 
     <RelativeLayout
