+++ app/src/main/res/xml/pref_headers.xml
@@ -3,17 +3,17 @@
     <!-- These settings headers are only used on tablets. -->
 
     <header
-        android:fragment="uk.ac.tees.donut.squad.SettingsActivity$GeneralPreferenceFragment"
+        android:fragment="uk.ac.tees.donut.squad.activities.SettingsActivity$GeneralPreferenceFragment"
         android:icon="@drawable/ic_info_black_24dp"
         android:title="@string/pref_header_general" />
 
     <header
-        android:fragment="uk.ac.tees.donut.squad.SettingsActivity$NotificationPreferenceFragment"
+        android:fragment="uk.ac.tees.donut.squad.activities.SettingsActivity$NotificationPreferenceFragment"
         android:icon="@drawable/ic_notifications_black_24dp"
         android:title="@string/pref_header_notifications" />
 
     <header
-        android:fragment="uk.ac.tees.donut.squad.SettingsActivity$DataSyncPreferenceFragment"
+        android:fragment="uk.ac.tees.donut.squad.activities.SettingsActivity$DataSyncPreferenceFragment"
         android:icon="@drawable/ic_sync_black_24dp"
         android:title="@string/pref_header_data_sync" />
